# src/app/Models/Menu.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Factories\HasFactory;

class Menu extends Model
{
    use HasFactory;

    protected $fillable = ['name', 'description', 'price', 'available_date'];

    public function orders()
    {
        return $this->hasMany(Order::class);
    }
}